moving plot function to plot_utils and updating the code accordingly
